cambios a stockConsola y productos agregar

+++ maderas/database/migrations/2020_08_12_165532_create_stock.php
@@ -17,7 +17,7 @@ class CreateStock extends Migration
             $table->bigIncrements('PK_stock');
             $table->unsignedBigInteger('id_PG_branches')->unsigned();
             $table->unsignedBigInteger('id_product')->unsigned();
-            $table->unsignedTinyInteger('avaible')->default(1);
+            $table->bigInteger('avaible')->default(1);
             $table->unsignedTinyInteger('status')->default(0);
             $table->timestamps();
